Rename elapsedTime statistics (mean, min, max, perc) to sampleTime to avoid confusion with existing elapsedTime function
sampleTime metrics are focused in sample time metrics, on the other hand elapsedTime is focused on the amount of time spent in some particular testplan/transaction/sampler by just getting the time between the element started and ended. This time (elapsedTime) does not only take into consideration sampling times, but also time processing pre&post processors, timers time, concurrency of thread groups, etc.

This is a non backwards compatible change, but we include it as is since we are in 0 major version, which states that the API is not stable, and we don't see much benefit right now on deprecating existing methods before removing them.

Additionally, improve documentation of statistics to clarify their purpose.