Avoid failing test plan when visualizer is used with BlazeMeterEngine
To avoid the failure we changed the model giving the engine the responsibility to show the visualizers (if supported). Resulting code is simpler and shorter.

Additionally fix some styling issues, when showInGui or similar method is used,

Finally, use EXIT_ON_CLOSE in created frame to avoid leaving the Swing threads running. On the other hand we use dispose when there are close listeners to avoid stopping thread unintentionally when visualizers are closed.