Add jmx2dsl conversion for jdbc elements
Additionally, improve Duration conversion to use the shortest and most meaningful conversion and relax MethodCall & Param imports methods to allow for more flexibility (ie: avoid having to have the class loaded in classpath to be able to include it's import). This later change is particularly helpful for jdbc Driver class specification, but might come in handy in the future.